Implemented batch and pubsub error fixed
