+++ sys_arc/__main__.py
@@ -9,7 +9,7 @@ _iam_setup.setup_iam()
 _iam_setup.pub_sub()
 _storage.setup_storage()
 _storage.data_in()
-_stream_pipe.run()
+# _stream_pipe.run()
